fix crashes on some systems, caused by combining libs kernel32 and msvcrt
This fixes #6 by partly reverting a commit @tonycoz contributed in #4:
"correctly pass linker options to MSVC"
74a1a8b6aad91340629f228a5faa313157a0a4c5

Notably on ActivePerl, using the ActiveState-provided MinGW the changes
introduced by that commit cause the programs compiled during testing to be
linked to both kernel32 and msvcrt, which causes them to crash when being
executed.

This commit fixes that by taking perllibs config variable out of the list
of options to be passed to the linker. This may cause some warnings on
systems that use MSVC instead of MinGW, but is highly preferred to crashes
and test failures.