Add analyze_binary feature
Sometimes it is useful to be able to inspect the binaries created by
Devel::CheckLib in order to, for instance, check its dependencies or
for the presence of some symbol or even some string.

This patch adds support for passing a custom callback to assert_libs
and its wrappers that is called back with the binary path in order to
do that kind of tasks.