Fix order of flags for gcc
In gcc order of libraries, object files and source code files is important.
If file A uses symbols from file B, then A must be specified prior to B.

If parameters are in incorrect order then gcc can throw fatal error
"undefined reference to...".

This patch fixes this problem. First is source file being compiled, next is
library specified for linking and after that are ld flags which may contain
additional libraries on which either source file or library specified for
linking depends.