Use rpath when compiling on macOS.
This fixes issue #337 for DBD::mysql. When compiling on macOS the rpath
must be explicitly set with linker option -Wl,-rpath or else the dynamic
loader will refuse to load the library. Note that many workarounds exists
like setting DYLD_LIBRARY_PATH or using install_name_tool. However, setting
rpath for the generated executable is to be preferred and is the
simplest solution.